Refactor initialization and discovery code
* move wipe_blockdevs() into blockdevmgr module
* split backstore/devices module into devices and initialize modules
* export fewer names from the backstore module and makes some test only
* move the identify module into the liminal module
* hoist the crypt and udev modules submodules up to be submodules of
strat_engine
