+++ src/engine/strat_engine/crypt.rs
