+++ src/engine/strat_engine/liminal/setup.rs
@@ -334,8 +334,11 @@ mod tests {
 
     use crate::engine::{
         strat_engine::{
-            backstore::{Backstore, LuksInfo, StratisInfo},
-            liminal::device_info::{LInfo, LLuksInfo},
+            backstore::Backstore,
+            liminal::{
+                device_info::{LInfo, LLuksInfo},
+                identify::{LuksInfo, StratisInfo},
+            },
             metadata::{MDADataSize, StratisIdentifiers},
             serde_structs::Recordable,
             tests::{crypt, loopbacked, real},
